pyproject.toml
[tool.poetry]
name = "sky"
version = "0.1.0"
description = ""
authors = ["antonglazunov <tagl91@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
openmeteo-requests = "^1.3.0"
requests-cache = "^1.2.1"
retry-requests = "^2.0.0"
numpy = "^2.2.2"
pandas = "^2.2.3"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
